chore: rename swagger-codegen
