+++ apps/web/app/(with-layout)/(home)/components/home-section-youtube.tsx
@@ -1,5 +1,5 @@
 import { apiRequest } from 'app/api/apiRequest';
-import { ApiResponseListTrendsYoutubeResponse } from 'typescript-swagger-codegen/data-contracts';
+import { ApiResponseListTrendsYoutubeResponse } from 'swagger-codegen/data-contracts';
 import { convertToEmbedUrl, validateYoutubeVideoList } from 'utils/youtube';
 
 export default async function HomeSectionYouTube() {
