+++ apps/web/app/(with-layout)/product-detail/[productId]/@modal/(.)write-review/hooks/presignedUrlApi.ts
@@ -4,7 +4,7 @@ import {
   ApiResponseReviewReceiptResponse,
   ReviewMediaRequest,
   ReviewReceiptRequest,
-} from 'typescript-swagger-codegen/data-contracts';
+} from 'swagger-codegen/data-contracts';
 
 export const getMediaPresignedUrl = async (request: ReviewMediaRequest) => {
   return await apiRequest<ApiResponseReviewMediaResponse>({
