+++ apps/web/app/review-modal/[reviewId]/detail/image/page.client.tsx
@@ -15,7 +15,7 @@ import LoadingSvg from 'components/loading/loading-svg';
 import {
   ApiResponseImageReviewDetailResponse,
   ImageReviewDetailResponse,
-} from 'typescript-swagger-codegen/data-contracts';
+} from 'swagger-codegen/data-contracts';
 
 const formatDateToJapanese = (dateString: string): string => {
   const date = new Date(dateString);
